feat(ansi): allow go 1.18
