+++ ansi/go.mod
@@ -1,6 +1,6 @@
 module github.com/charmbracelet/x/ansi
 
-go 1.21
+go 1.18
 
 require (
 	github.com/lucasb-eyer/go-colorful v1.2.0
