+++ ansi/util.go
@@ -90,3 +90,17 @@ func XParseColor(s string) color.Color {
 	}
 	return nil
 }
+
+type ordered interface {
+	~int | ~int8 | ~int16 | ~int32 | ~int64 |
+		~uint | ~uint8 | ~uint16 | ~uint32 | ~uint64 | ~uintptr |
+		~float32 | ~float64 |
+		~string
+}
+
+func max[T ordered](a, b T) T { //nolint:predeclared
+	if a > b {
+		return a
+	}
+	return b
+}
